% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_output.R
\name{nsim_output_download}
\alias{nsim_output_download}
\title{处理结果的输出方式}
\usage{
nsim_output_download(brand = "JBLH", data, each_page = 7000L,
  skip_row = 2L)
}
\arguments{
\item{brand}{品牌}

\item{data}{传入数据}

\item{each_page}{每页数}

\item{skip_row}{行数}

\item{file_name}{文件名}
}
\value{
返回值
}
\description{
处理结果的输出方式
}
\examples{
nsim_output_download();
}
