% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_output.R
\name{nsim_output_download_nscs}
\alias{nsim_output_download_nscs}
\title{处理结果最终结构}
\usage{
nsim_output_download_nscs(brand = "JBLH", each_page = 7000L,
  skip_row = 2L)
}
\arguments{
\item{brand}{品牌}

\item{each_page}{页数}

\item{skip_row}{忽略行}
}
\value{
返回值
}
\description{
处理结果最终结构
}
\examples{
nsim_output_download_nscs();
}
