% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_output.R
\name{nsim_output_download_test}
\alias{nsim_output_download_test}
\title{处理结果的输出方式}
\usage{
nsim_output_download_test(brand = "JBLH", each_page = 3000L,
  skip_row = 2L, file_name = "output.res.paging.20191023T.xlsx")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理结果的输出方式
}
\examples{
nsim_output_download();
}
